log-generator/requirements.txt
# log-generator/requirements.txt
Faker
kafka-python